requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.3
cachetools==3.1.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.61.0
grpcio==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic-core==2.4.0
pydeck==0.8.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.27.2
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uritemplate==3.0.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
xds-protos==1.59.0
zipp==3.17.0
